separate SpecoTest
